+++ CHANGES.md
@@ -154,6 +154,7 @@ Others:
 -   Disabled tenant-api & tenant-db when `enableMultiTenants` = false
 -   Excluded organisations that are owners of thesauruses (keyword taxonomies) from being considered as owners of datasets via CSW connector
 -   Fix data.json connector dcat-dataset-strings aspect so keywords are stored correctly
+-   Fix CSW connector may process XML response incorrectly and report `no id` error
 
 ## 0.0.55